+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## next
 
 - Improved error handling in data prepare handler
+- Changed location path in signature details popup to use `[index]` instead of `pick(index)`
 
 ## 1.0.0-beta.38 (19-05-2020)